requirements.txt
defusedxml==0.4.1
Django==1.10
django-allauth==0.26.1
djangorestframework==3.4.4
mysqlclient==1.3.7
oauthlib==1.1.2
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
python-memcached==1.58
git+https://github.com/mobolic/facebook-sdk.git#egg=facebook-sdk
celery